// src/components/Modals/BookHotelDrawer/BookHotelDrawer.type.ts
export type FilterInfoType = {
  startDate: string;
  endDate: string;
  numAdults: number;
  numKids: number;
  numRooms: number;
  City: any;
  minNightlyRateUsdCents: number;
  maxNightlyRateUsdCents: number;
  starsArr: number[];
};

export type HotelImageType = {
  caption: string;
  imageThumbnailUrl: string;
  imageUrl: string;
  sourceImageUrl: string;
};